refactor(webapi): move Ban entity into Athlete aggregate
Ban has no independent lifecycle outside the Athlete aggregate.
Move Ban, BanErrors, and BanConfiguration from Features/Bans/ into
Features/Athletes/, update all namespace references across migrations,
model snapshot, tests, and builders.

Closes #385
