+++ src/KRAFT.Results.WebApi/Features/Athletes/Ban.cs
@@ -1,7 +1,6 @@
-﻿using KRAFT.Results.WebApi.Abstractions;
-using KRAFT.Results.WebApi.Features.Athletes;
+using KRAFT.Results.WebApi.Abstractions;
 
-namespace KRAFT.Results.WebApi.Features.Bans;
+namespace KRAFT.Results.WebApi.Features.Athletes;
 
 internal sealed class Ban
 {
